Statdist methods now in StatDistResult
